requirements.txt
# Dependencies for Devin API integration
requests==2.31.0
python-dotenv==1.0.0
prefect==3.6.2
